Refactor out reducer from Player component
