+++ client/app/reducers/editor.ts
